+++ modules/decrypt-node/src/decrypt_client.ts
@@ -3,7 +3,12 @@
 
 import { _decryptStream } from './decrypt_stream'
 import { _decrypt } from './decrypt'
-import { CommitmentPolicy, needs } from '@aws-crypto/material-management-node'
+import {
+  CommitmentPolicy,
+  needs,
+  SignaturePolicy,
+  ClientOptions,
+} from '@aws-crypto/material-management-node'
 
 type CurryFirst<fn extends (...a: any[]) => any> = fn extends (
   _: any,
@@ -13,17 +18,58 @@ type CurryFirst<fn extends (...a: any[]) => any> = fn extends (
   : never
 
 export function buildDecrypt(
-  commitmentPolicy: CommitmentPolicy = CommitmentPolicy.REQUIRE_ENCRYPT_REQUIRE_DECRYPT
+  options: CommitmentPolicy | Partial<ClientOptions> = {}
 ): {
+  decryptUnsignedMessageStream: (
+    ...args: CurryFirst<typeof _decryptStream>
+  ) => ReturnType<typeof _decryptStream>
   decryptStream: (
     ...args: CurryFirst<typeof _decryptStream>
   ) => ReturnType<typeof _decryptStream>
   decrypt: (...args: CurryFirst<typeof _decrypt>) => ReturnType<typeof _decrypt>
 } {
+  const {
+    commitmentPolicy = CommitmentPolicy.REQUIRE_ENCRYPT_REQUIRE_DECRYPT,
+    maxEncryptedDataKeys = false,
+  } = typeof options === 'string' ? { commitmentPolicy: options } : options
+
   /* Precondition: node buildDecrypt needs a valid commitmentPolicy. */
   needs(CommitmentPolicy[commitmentPolicy], 'Invalid commitment policy.')
+  /* Precondition: node buildDecrypt needs a valid maxEncryptedDataKeys. */
+  needs(
+    maxEncryptedDataKeys === false || maxEncryptedDataKeys >= 1,
+    'Invalid maxEncryptedDataKeys value.'
+  )
+
+  const clientOptions: ClientOptions = {
+    commitmentPolicy,
+    maxEncryptedDataKeys,
+  }
   return {
-    decryptStream: _decryptStream.bind({}, commitmentPolicy),
-    decrypt: _decrypt.bind({}, commitmentPolicy),
+    decryptUnsignedMessageStream: _decryptStream.bind(
+      {},
+      {
+        signaturePolicy: SignaturePolicy.ALLOW_ENCRYPT_FORBID_DECRYPT,
+        clientOptions,
+      }
+    ),
+    decryptStream: _decryptStream.bind(
+      {},
+      {
+        signaturePolicy: SignaturePolicy.ALLOW_ENCRYPT_ALLOW_DECRYPT,
+        clientOptions,
+      }
+    ),
+    decrypt: _decrypt.bind(
+      {},
+      {
+        signaturePolicy: SignaturePolicy.ALLOW_ENCRYPT_ALLOW_DECRYPT,
+        clientOptions,
+      }
+    ),
   }
 }
+
+// @ts-ignore
+const { decryptUnsignedMessageStream, decryptStream, decrypt } = buildDecrypt()
+decryptUnsignedMessageStream({} as any)